api/pyproject.toml
[project]
name = "AlgoAlchemy"
version = "0.1.0"
description = "A clever mix of algorithms and the magic (alchemy) of Python, implemented with both Flask and FastAPI."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=37.1.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "ruff>=0.11.8",
]
